move api's type.py to core_python
